Another lamire

+++ src/models/Product.java
@@ -5,31 +5,31 @@ import java.util.Objects;
 
 public class Product implements Serializable {
     private final String name;
-    private int originalPrice = 0;
-    private int marketPrice;
+    private float originalPrice = 0;
+    private float marketPrice;
 
     public Product(String name) {
         this.name = name;
     }
 
-    public Product(String name, int originalPrice) {
+    public Product(String name, float originalPrice) {
         this(name);
         this.originalPrice = originalPrice;
     }
 
-    public int getMarketPrice() {
+    public float getMarketPrice() {
         return marketPrice;
     }
 
-    public void setMarketPrice(int marketPrice) {
+    public void setMarketPrice(float marketPrice) {
         this.marketPrice = marketPrice;
     }
 
-    public void setOriginalPrice(int originalPrice) {
+    public void setOriginalPrice(float originalPrice) {
         this.originalPrice = originalPrice;
     }
     
-    public int getOriginalPrice() {
+    public float getOriginalPrice() {
         return originalPrice;
     }
 
@@ -41,9 +41,7 @@ public class Product implements Serializable {
     public String toString(){
         if(getOriginalPrice() == 0)
             return getName();
-        // if(getMarketPrice() == 0)
         return getName() + ":" + getOriginalPrice();
-        // return getName() + ":" + getOriginalPrice() + ":" + getMarketPrice();
     }
 
     @Override
